backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
psycopg2-binary==2.9.9
pillow==10.1.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
accelerate==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3